backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
mysql-connector-python==9.4.0
Werkzeug==3.1.3
PyJWT==2.10.1
psycopg2-binary==2.9.11